requirements.txt
ffmpeg==1.4
ffmpeg_python==0.2.0
joblib==1.4.2
numpy==2.2.6
openai_whisper==20240930
pandas==2.2.3
torch==2.5.1+cu121
